Stimulator refactoring
